+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3599,7 +3599,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KKEEKFEP1KAY3BTGES1BV0HV"),
+  id: z.string().default("01KKGBFE8QFVDY7KEKRT7B2ZWW"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),